requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-huggingface>=0.1.0
chromadb>=0.5.0
playwright>=1.40.0
fastapi>=0.100.0
uvicorn>=0.24.0
streamlit>=1.30.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.5.0